Fixed missing saves
Made FileManager check for a missing saves folder.
Will create a saves folder in writeFile if not found.

+++ src/server/api/save-manager.ts
@@ -135,7 +135,7 @@ export class SaveManager {
  *
  */
 export class FileManager {
-    private static FilePath = "src/server/saves";
+    private static FilePath = "saves";
     private static fs = require("fs");
 
     /*
@@ -150,6 +150,9 @@ export class FileManager {
      */
     public static writeFile(saveName: string, saveContents: iSave) {
         try {
+            if (!this.fs.existsSync(this.FilePath)) {
+                this.fs.mkdirSync(this.FilePath, { recursive: true });
+            }
             this.fs.writeFileSync(
                 this.FilePath + "/" + saveName,
                 JSON.stringify(saveContents),
@@ -171,6 +174,7 @@ export class FileManager {
      * Output: iSave
      */
     public static loadFile(fileName: string): iSave {
+        if (!this.fs.existsSync(this.FilePath)) return {} as iSave;
         const a = this.fs.readFileSync(this.FilePath + "/" + fileName);
         if (a) return JSON.parse(a);
         return {} as iSave;
@@ -184,6 +188,7 @@ export class FileManager {
      * Output: Array of file names
      */
     public static getFileNames(): string[] {
+        if (!this.fs.existsSync(this.FilePath)) return [];
         const fileNames: string[] = [];
         const files = this.fs.readdirSync(this.FilePath);
 
@@ -202,6 +207,7 @@ export class FileManager {
      * Input: host id, client id
      */
     public static deleteFile(fileName: string) {
+        if (!this.fs.existsSync(this.FilePath)) return [];
         this.fs.unlink(this.FilePath + "/" + fileName, (err) => {
             if (err) {
                 console.error(err);
